+++ tests/onnx2c_refs/test_onnx10_hello.c
@@ -1,4 +1,5 @@
 #include <stdint.h>
+#include <sys/types.h>
 #include <stdbool.h>
 #include "ops_scalar_f32.h"
 
@@ -67,10 +68,10 @@ static const float weight_MatMul_dense_4_bias[1] = {
 };
 
 void node1_linear_f32(const float input[1][1], const float weight[16][1], const float bias[16], float out[1][16]) {
-    for (int64_t i0 = 0; i0 < 1; ++i0) {
-        for (int64_t j = 0; j < 16; ++j) {
+    for (ssize_t i0 = 0; i0 < 1; ++i0) {
+        for (ssize_t j = 0; j < 16; ++j) {
             float acc = 0.0f;
-            for (int64_t t = 0; t < 1; ++t) {
+            for (ssize_t t = 0; t < 1; ++t) {
                 acc += input[i0][t] * weight[j][t];
             }
             out[i0][j] = acc + bias[j];
@@ -79,18 +80,18 @@ void node1_linear_f32(const float input[1][1], const float weight[16][1], const 
 }
 
 void node2_relu_f32(const float a[1][16], float out[1][16]) {
-    for (int64_t i0 = 0; i0 < 1; ++i0) {
-        for (int64_t i1 = 0; i1 < 16; ++i1) {
+    for (ssize_t i0 = 0; i0 < 1; ++i0) {
+        for (ssize_t i1 = 0; i1 < 16; ++i1) {
             out[i0][i1] = ref_scalar_f32_relu(a[i0][i1]);
         }
     }
 }
 
 void node3_linear_f32(const float input[1][16], const float weight[16][16], const float bias[16], float out[1][16]) {
-    for (int64_t i0 = 0; i0 < 1; ++i0) {
-        for (int64_t j = 0; j < 16; ++j) {
+    for (ssize_t i0 = 0; i0 < 1; ++i0) {
+        for (ssize_t j = 0; j < 16; ++j) {
             float acc = 0.0f;
-            for (int64_t t = 0; t < 16; ++t) {
+            for (ssize_t t = 0; t < 16; ++t) {
                 acc += input[i0][t] * ((float*)weight)[j * 1 + t * 16];
             }
             out[i0][j] = acc + bias[j];
@@ -99,18 +100,18 @@ void node3_linear_f32(const float input[1][16], const float weight[16][16], cons
 }
 
 void node4_relu_f32(const float a[1][16], float out[1][16]) {
-    for (int64_t i0 = 0; i0 < 1; ++i0) {
-        for (int64_t i1 = 0; i1 < 16; ++i1) {
+    for (ssize_t i0 = 0; i0 < 1; ++i0) {
+        for (ssize_t i1 = 0; i1 < 16; ++i1) {
             out[i0][i1] = ref_scalar_f32_relu(a[i0][i1]);
         }
     }
 }
 
 void node5_linear_f32(const float input[1][16], const float weight[1][16], const float bias[1], float out[1][1]) {
-    for (int64_t i0 = 0; i0 < 1; ++i0) {
-        for (int64_t j = 0; j < 1; ++j) {
+    for (ssize_t i0 = 0; i0 < 1; ++i0) {
+        for (ssize_t j = 0; j < 1; ++j) {
             float acc = 0.0f;
-            for (int64_t t = 0; t < 16; ++t) {
+            for (ssize_t t = 0; t < 16; ++t) {
                 acc += input[i0][t] * weight[j][t];
             }
             out[i0][j] = acc + bias[j];
